fix(img_tools): wrong dtype in row_zoom_img call

+++ pandora2d/img_tools.py
@@ -513,7 +513,7 @@ def shift_subpix_img(img_right: xr.Dataset, subpix: int, row: bool = True, order
         # To avoid propagation of nan in shifted data, we use -9999 instead of nan if necessary
         no_data = convert_no_data(img_right.attrs["no_data_img"])
 
-        for ind in np.arange(1, subpix):
+        for ind in np.arange(1, subpix, dtype=int):
             if row:
                 img_right_shift.append(
                     row_zoom_img(
